@@ -1,83 +1,64 @@
 ﻿using System;
 using System.Threading.Tasks;
-using Grpc.Net.Client;
 using Trinsic.Services.Common.V1;
+using Trinsic.Services.UniversalWallet.V1;
 
-namespace Trinsic
+namespace Trinsic;
+
+public class ProviderService : ServiceBase
 {
-    public class ProviderService : ServiceBase
+    public ProviderService(WalletProfile walletProfile, ServerConfig? serverConfig) : base(walletProfile, serverConfig)
     {
-        public ProviderService()
-            : this(new ServerConfig
-            {
-                Endpoint = "prod.trinsic.cloud",
-                Port = 443,
-                UseTls = true
-            })
-        {
+        Client = new Provider.ProviderClient(Channel);
+    }
 
-        }
-        public ProviderService(ServerConfig config)
-            : this(ServiceBase.CreateChannelIfNeeded($"{(config.UseTls ? "https" : "http")}://{config.Endpoint}:{config.Port}"))
-        {
-        }
+    public Provider.ProviderClient Client { get; }
 
-        public ProviderService(GrpcChannel channel)
+    /// <summary>
+    /// Initates the participant onboarding flow using the input contact method
+    /// </summary>
+    /// <param name="request"></param>
+    /// <returns></returns>
+    public async Task<InviteResponse> InviteParticipant(InviteRequest request)
+    {
+        if (request.ContactMethodCase == InviteRequest.ContactMethodOneofCase.None)
         {
-            // We must store a reference to the channel, otherwise it gets collected
-            Channel = channel;
-            ProviderClient = new Provider.ProviderClient(Channel);
+            throw new Exception("Contact method must be set.");
         }
 
-        public GrpcChannel Channel { get; }
-        public Provider.ProviderClient ProviderClient { get; }
-
-        /// <summary>
-        /// Initates the participant onboarding flow using the input contact method
-        /// </summary>
-        /// <param name="request"></param>
-        /// <returns></returns>
-        public async Task<InviteResponse> InviteParticipant(InviteRequest request)
+        try
         {
-            if (request.ContactMethodCase == InviteRequest.ContactMethodOneofCase.None)
-            {
-                throw new Exception("Contact method must be set.");
-            }
+            var response = await Client.InviteAsync(request, BuildMetadata(request));
+            return response;
+        }
+        catch (Exception e)
+        {
+            Console.WriteLine(e);
+            throw;
+        }
+    }
 
-            try
-            {
-                var response = await ProviderClient.InviteAsync(request, GetMetadata(request));
-                return response;
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine(e);
-                throw;
-            }
+    /// <summary>
+    /// Gets the status of the onboarding invitation for a given invitation ID
+    /// </summary>
+    /// <param name="request"></param>
+    /// <returns></returns>
+    public async Task<InvitationStatusResponse> InvitationStatus(InvitationStatusRequest request)
+    {
+        if (string.IsNullOrWhiteSpace(request.InvitationId))
+        {
+            throw new Exception("Onboarding reference ID must be set.");
         }
 
-        /// <summary>
-        /// Gets the status of the onboarding invitation for a given invitation ID
-        /// </summary>
-        /// <param name="request"></param>
-        /// <returns></returns>
-        public async Task<InvitationStatusResponse> InvitationStatus(InvitationStatusRequest request)
+        try
         {
-            if (string.IsNullOrWhiteSpace(request.InvitationId))
-            {
-                throw new Exception("Onboarding reference ID must be set.");
-            }
-
-            try
-            {
-                var response = await ProviderClient.InvitationStatusAsync(request, GetMetadata(request));
-                return response;
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine(e);
-                throw;
-            }
+            var response = await Client.InvitationStatusAsync(request, BuildMetadata(request));
+            return response;
+        }
+        catch (Exception e)
+        {
+            Console.WriteLine(e);
+            throw;
         }
     }
 }
